4.17. Шаблонизируй то
