+++ src/const.js
@@ -18,4 +18,11 @@ const DAY_FORMAT = 'MMM DD';
 const POINT_TIME_FORMAT = 'HH:mm';
 const NEW_POINT_TIME_FORMAT = 'DD/MM/YY HH:mm';
 
-export { SORT_TYPES, POINT_TYPES, DESTINATIONS_DESCRIPTIONS, DAY_FORMAT, POINT_TIME_FORMAT, NEW_POINT_TIME_FORMAT };
+const FilterType = {
+  EVERYTHING: 'everything',
+  FUTURE: 'future',
+  PRESENT: 'present',
+  PAST: 'past',
+};
+
+export { SORT_TYPES, POINT_TYPES, DESTINATIONS_DESCRIPTIONS, DAY_FORMAT, POINT_TIME_FORMAT, NEW_POINT_TIME_FORMAT, FilterType };
